requirements.txt
lightning==2.5.1
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
rich==14.0.0
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
